backend/python_backend/requirements.txt
# Existing requirements
flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
scikit-learn==1.3.0
numpy==1.24.3
Pillow==10.0.1
pygltflib==1.16.1

# New requirements for 3D thumbnail generation
trimesh==3.23.5
pyrender==0.1.45
requests==2.31.0

# Optional: for better 3D model support
meshio==5.3.4
